Wholesalerdashboard

+++ database/seeds/UsersSeeder.php
@@ -13,37 +13,54 @@ class UsersSeeder extends Seeder
      */
     public function run()
     {
+        $connection = config('database.default');
+        $driver = config("database.connections.{$connection}.driver");
+
         $faker = Factory::create();
         //reset the table
-        DB::statement("SET FOREIGN_KEY_CHECKS=0");
-        DB::table('users')->truncate();
+        if($driver == 'mysql') {
+            DB::statement("SET FOREIGN_KEY_CHECKS=0");
+            DB::table('users')->truncate();
+        }
 
         DB::table('users')->insert([
                 [
-                    'name' => "Admin Jack",
-                    "slug" =>"admin-doe",
-                    'email' => "admin@blog.com",
+                    'name' => "Admin Nnuro",
+                    'firstname' => "Admin",
+                    'lastname' => "Nnuro",
+                    "slug" =>"admin-nnuro",
+                    "type" =>"admin",
+                    'email' => "admin@nnuro.com",
                         'password' => bcrypt('secret'),
                         "bio" => $faker->text(rand(250, 300))
                 ],
                 [
-                    'name' => "John Doe",
-                    "slug" =>"john-doe",
-                    'email' => "johndoe@test.com",
-                        'password' => bcrypt('secret'),
-                        "bio" => $faker->text(rand(250, 300))
-                ],
-                [
-                    'name' => "Jane Doe",
-                    "slug" =>"jane-doe",
-                    'email' => "janedoe@test.com",
+                    'name' => "Wholesaler Jane",
+                    "slug" =>"wholesaler-jane",
+                    'firstname' => "Jane",
+                    'lastname' => "Doe",
+                    "type" =>"wholesaler",
+                    'email' => "wholesaler@nnuro.com",
                     'password' => bcrypt('secret'),
                     "bio" => $faker->text(rand(250, 300))
                 ],
                 [
-                'name' => "Mark Otoo",
-                "slug" =>"mark-otoo",
-                'email' => "markotoo@test.com",
+                    'name' => "Retailer John",
+                    "slug" =>"retailer-john",
+                    'firstname' => "Retailer",
+                    'lastname' => "John",
+                    "type" =>"retailer",
+                    'email' => "retailer@nnuro.com",
+                        'password' => bcrypt('secret'),
+                        "bio" => $faker->text(rand(250, 300))
+                ],
+                [
+                'name' => "Consumer Mark",
+                "slug" =>"consumer-mark",
+                'firstname' => "Mark",
+                'lastname' => "Consumer",
+                "type" =>"customer",
+                'email' => "consumer@nnuro.com",
                 'password' => bcrypt('secret'),
                 "bio" => $faker->text(rand(250, 300))
                 ]
